Extract pattern types.
I extracted the pattern types in order to be able to break apart some of
its section in separate modules. The problem with having the data and
functions in the same module is that it makes it impossible to split the
functions in separated modules due to their dependencies on the data types
(thus creating a circular module dependency, which Haskell frowns upon).