Replace patterns with type variables.
Even though this change seems to be quite large only two
small changes have been done that ripple through all the
sources.

1. Change the data types to begin abstracting away patterns.

We have two patterns: Pattern and BoundPattern that should be
abstracted away in order to properly decouple the code. This
first change removed by-name references to patterns from the
data types, but still keeps then in the rest of the code.

2. Remove the Constructor reference from the Pattern constructor.

The Constructor data type also includes Map/Set/List constructors,
and the corresponding Pattern data constructor is always instantiated
with Symbols or Literals.

This change has two benefits:

a. It's more type-safe since now by construction you can only use symbols
or literals with the Pattern data constructor.
b. The code is simplified since we do not need to add a variable for the
BoundPattern that appears in Constructors to the Pattern data type. This
would mean that we will have two fix points in the same data structure.