centralize all information regarding block headers
All information relating to block headers is now either in
CMakeLists.txt or include/runtime/header.h depending on whether it
is specific to C or if it is used by both C and LLVM.

We also reduce the number of layout bits from 16 to 10 and the number
of string length bits from 45 to 40 to allow a total of 10 bits in each
token to be used to store the sort of the token. This will be
accomplished in the next PR, however.