Parser: update rules to reflect KORE syntax change for `ensures`
There was an [issue](https://github.com/kframework/k/issues/1945) where
the the rule
```
rule LEFT => RIGHT requires REQUIRES ensures ENSURES
```
is translated to
```
axiom \implies(REQUIRES, \and(\equals(LEFT, RIGHT), ENSURES))
```
so that if ENSURES is \bottom and REQUIRES is anything else, then the axiom is axiom \bottom which is unsound. Actually, we want to say
```
axiom \implies(REQUIRES, \equals(LEFT, \and(RIGHT, ENSURES)))
```

There is a [pull request](https://github.com/kframework/k/pull/2061)
that changes the K frontend to emit the correct KORE syntax. This commit
implements the change in the KORE syntax for the llvm-backend parsers.

Fixes #417

Signed-off-by: Erik Kaneda <erik.kaneda@runtimeverification.com>