Parser: fix ensures parsing rule and update tests (#421)
This change fixes commit 19e3019 which contained an incorrect pattern
match. This was not caught until the frontend CI ran the pl-tutorial
tests with an updated AST from the KORE emitter. Update relevant tests
to exercise the new AST format so failures like this can be caught in
the llvm backend tests.

Signed-off-by: Erik Kaneda <erik.kaneda@runtimeverification.com>