Add documentation for local macOS builds (#422)
This documentation mirrors the instructions for Ubuntu 20.04 that are
also given in `INSTALL.md`. The environment is configured using the
system package manager (rather than Nix, as is the case in the CI
environment).

I have verified the instructions on a clean VM install of macos 11, as
well as my own development system.

Running the unit tests via `./ciscript` creates DWARF symbol files
(`.dSYM`) on macos; this PR adds these to the git ignore list.