Add hook for kompiled directory metadata (#425)
* Explicitly disable clang-format

I use an editor hook to run clang-format automatically - adding this
line ensures that I don't accidentally re-enable it on this project and
format something I didn't mean to. If we have an actual house style I'm
happy to write up a formatter for it.

* Always add global kompiled dir symbol

Previously this symbol was only added when debugging was enabled; this
commit makes it unconditional (and pushes the debug check inwards to the
point where symbol information is initialised).

* Generate hooked value

This commit uses the existing hook machinery to load the global variable
containing the kompiled directory, then box it into a K string value.

* Rework hook to C++ implementation

This is instead of using code generation, which is easier to test and
maintain.

* Unit test for kompiledDir hook

* Use memcmp rather than strcmp

* Explicit length check