Update `k-rule-find` to handle multiple rule labels (#829)
Fixes #725

This PR updates the `k-rule-find` tool to return multiple rule labels
when the given location has `cool` and `heat` rules, like when using
`seqstrict` and `strict`.

This PR will break the
[`context-labels`](https://github.com/runtimeverification/k/tree/master/k-distribution/tests/regression-new/context-labels)
test on the K frontend as the output expected there has only one rule
when more than one will be returned after this modification.