Replacing date/gdate with perl program (#833)
Fixes https://github.com/runtimeverification/k/issues/3610
This PR replaces the use of `date` and `gdate`, depending on the user
system, to get the current timestamp in milliseconds with a simple
`Perl` program to improve compatibility across Unix-like systems.
 
This modification is needed due to the following:
- `date` in MacOS is a FreeBSD version that doesn't support (return)
timestamps in milliseconds as we need in `llvm-kompile --profile`
introduced by #793.
- `gdate` isn't a default package nor a standalone package to be
installed in MacOS. Instead, it's commonly installed with `coreutils`.
We do not want to ask our users to install all `coreutils` when we only
need `gdate` from it and to use it in an optional flag.
- Other options were considered as `python time` and `bash
${EPOCHREALTIME}`, but this was the simplest to use and with common
default dependencies in Unix-like systems.