Informal LLVM 17 compatibility (#856)
As expected, the changes required to support LLVM 17 are extremely
minimal; only one API that we use changes from version 16. This PR adds
initial support for LLVM 17, but does not yet add it as a fully-tested
version as Nix has not yet added upstream support.

Because the changes here are minimal and almost entirely hidden behind
conditional compilation, my preference is to merge this PR before Nix
support for LLVM 17 lands. Existing CI and supported systems will not be
affected, and it allows macOS users (and other rolling releases) to use
the latest version of LLVM.