Modify C++ code in llvm backend to use multi-ary \and and \or (#844)
Blocked on
https://github.com/runtimeverification/haskell-backend/pull/3666

There's only a few places in the C++ code in the llvm backend that need
to be changed. We need to not emit `\left-assoc{}(\or{}())`, and we need
to fix kprint when it receives a configuration that has a multi-ary \and
and \or.

We also revert the temporary changes in the last llvm-backend PR.

Total list of changes:

* Revert temporary changes to scala code from previous PR
* Write code that converts multi-ary \and/\or to binary \and/\or prior
to unparsing in kprint.
* Use multi-ary \or when writing search results to output.
* Fix test output files.
* Make KORE parser for patterns return a shared_ptr instead of a
unique_ptr. This is because we need to be able to return the child of
something that was parsed if we parse an \and/\or with one child. By
that point, the term is stored in a shared_ptr and we cannot get it back
to being a unique_ptr, so we need to return a shared_ptr everywhere in
this case.
* Fix some code in the python and C bindings that broke as a result of
the previous change.
* Modify KORE parser to fix AST generated by \and or \or with 0 or 1
child.

---------

Co-authored-by: rv-jenkins <devops@runtimeverification.com>