Fix type printing in LLVM >= 16 (#861)
This PR fixes
https://github.com/runtimeverification/llvm-backend/issues/810, the last
outstanding bit of migration that needed to happen for us to support
LLVM 16+ opaque pointers. The full context for this change is in the
original issue.

To fix the issue, we need to delay the conversion from `ValueType` to
`llvm::Type *` until a deeper point in the call stack when emitting
calls to `addMatchFunction`. The changes here are designed to minimise
the surface area of modified code; we can get away without changing the
`MatchLog` data structure or the `printValueOfType` function by just
converting `ValueType`s to an _imitation_ of the legacy LLVM type
strings. We might want to revisit this architecture in the future, but
for now I'm happy with the design.

The changes made here are tested by the tests for `k-rule-apply`; on
platforms with LLVM 16+ we had marked these tests as expectedly failing,
and we just remove that label here. In the second commit you can see CI
failing because the `XFAIL` is incorrect.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>