Inline invocations of `llc` for main module (#814)
This follows #794 in the pursuit of improving the performance of the
backend's compilation pipeline; the philosophy of the change is
identical (remove places where we materialize code in the compilation
pipeline, and instead use the LLVM libraries to do it all in-memory).

There are a few more changes required in this PR in order to expose the
full generality of the options we previously supported for `llc`, but
these are largely boilerplate.

The default behaviour of `llvm-kompile` is now to use the new pipeline,
but it's possible to opt back out of it for debugging.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>