Refactor cmake-configured C++ code (#870)
This PR follows up on #869 to refactor our use of CMake to configure C++
code with build-specific information. Instead of configuring individual
files, we push all `@...@` stanzas into `config/macros.h` and use that
file as the single source of configured literals.

There is a small change to the implementation of the LLVM header in
`CreateTerm.h` to support this.