Use domain value argument sort if possible (#875)
This PR implements a feature requested by @gtrepta whereby `getSort()`
calls on _domain value_ patterns that are part of a definition will
return the first formal sort argument (e.g. `SortX{}` in
`\dv{SortX{}}("...")` rather than `nullptr`. Doing so simplifies
printing and serialization logic in some parts of the backend.

As well as CI passing, I have tested this on upstream K and it does not
appear to break anything; my suspicion is that we just weren't relying
on these sorts being `nullptr` previously.

The PR also pins a specific version of `clang-format`, which we weren't
doing previously.