Refactoring proof trace code (#876)
As discussed previously, this PR is an initial cleanup and refactoring
effort for the existing proof-hint generation code implemented for
rewrite, function and hook events. The changes in this PR don't add any
new features to the traces; they just reorganise the existing code.

Most of the changes made are documentation and identifying duplicated
code across different call-sites that can be merged together.

I plan to port the reorganisation in #862 over to the infrastructure in
this PR once it is merged.

The trace format is not tested in the backend currently (this is future
work), but I have verified that a proof trace (that uses all event
types) generated using this branch is byte-for-byte identical to one
generated from the current master branch.