Flags to turn on/off proof hint instrumenation and output (#879)
This PR adds two flags to our scripts to turn the proof hint
instrumentation and output on and off:
- a flag `--proof-hint-instrumentation` is added to the `llvm-kompile`
script. When the flag is passed the generated llvm code is instrumented
with proof hint generation code.
- a flag `--proof-hints` is added to the `llvm-krun` script. When the
flag is passed to an instrumented interpreter binary, the binary will
output proof hints.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>