Disable iterative matching when proof hints are enabled (#887)
When proof hint generation is enabled, we want to be able to emit every
intermediate configuration into the trace; the current implementation
only does so when we are at a rewrite step that has not used iterative
compilation to merge two decision tree matrices. This PR disables the
use of iterative compilation and residuals when generating code with
trace instrumentation.

Doing so has no effect on the soundness of the generated code; iterative
compilation is a rarely-used optimization.

The first 3 commits in this PR are a slight refactoring of our usage of
command-line options when running code generation.