Reorganisation: extract C bindings to core library
This commit reorganises the existing C bindings for the LLVM backend by
extracting some of their AST manipulation code to the newly-created core
bindings library. It does not change the set of features offered by the
bindings; the diff does not look quite like a straight reorganisation
because the implementation changes from "pure C" style to direct usage
of the underlying C++ data structures, but the intent of the code
remains identical.