Refactor usages of `std::ostringstream` (#912)
A common pattern in the backend's code is to do the following:
```c++
std::ostringstream out;
pattern->print(out);
do_something(out.str());
```

In #911, we introduced a helper function `ast_to_string` that wraps this
common pattern up; this PR follows up that change by mechanically
refactoring the remainder of the codebase to use the helper. After the
change, the above code would be:
```c++
do_something(ast_to_string(*pattern));
```

Each commit in the PR (bar the first) applies the same conceptual
refactoring to a single file; no actual behavioural changes other than
in the first commit are implemented, and so this should be an easy PR to
review.

In future work, I'd like to further clean up this code and make better
use of `fmt` across the codebase, but for now this is a useful
self-contained change to make.