Print hooked sort elements correctly (#914)
This PR changes the way that we print and serialize collection elements.
Previously, we assumed that the elements of collections were uniformly
of sort `KItem`. This is indeed the case for the built-in `Map`, `Set`
and `List` sorts, but the assumption breaks down in the presence of
custom collection sorts that sort their elements differently (for
example, defining a custom `Map` that requires its keys to only be
integers). Terms of these custom sorts are printed incorrectly, with
spurious injections into `KItem` rather than their actual element sorts.

The solution to the problem is to look up the argument sorts of the
element constructor when printing and serializing collections; the
required information is already present in the binary and just requires
a bit of refactoring to access.

This PR makes the following changes:
* Adds @geo2a's minimised test case from #886 as a new integration test.
* Refactors the generation of argument and return sort tables when the
configuration parser is emitted; previously these did not use the
`emitDataTable` infrastructure and so required custom C functions to be
written to index into them.
* Look up element sorts using the newly-accessible table when printing
collections.

Fixes #886