Expose function evaluation to bindings (#905)
This PR adds a new feature to the backend's binding API that allows for
function symbols to be evaluated on a set of AST pattern arguments, with
the result returned as a pattern also. This is motivated by the proof
generation project; the execution trace format generated by the backend
to summarise a concrete execution can contain the labels and arguments
for function call events, but not the functions' return values without
breaking tail-call optimisation.

Most of the prerequisite work for this PR was done in #908 and #911; the
actual changes here are quite small:
* Adds a sort-aware term-to-pattern reification method that uses the new
`rawTerm{}` wrapper.
* Marshals arguments and returns between pattern and term level in the
core bindings API.
* Adds a test case demonstrating a few simple cases for this feature; as
ever the bulk of the testing ought to be performed in Pyk.