Enable clang-tidy performance checks (#932)
This follows up on #924 by enabling the performance set of `clang-tidy`
checks. As in the original PR, each individual commit following the
first one mechanically fixes simple issues related to a single check. I
suspect the changes here will be pretty uncontroversial; the bulk of the
diff is just ensuring that we're using const references everywhere it's
appropriate to do so.

I have left one check silenced (integer-to-pointer conversions) as we do
enough valid instances of these in the backend runtime that it would be
annoying to use manual suppressions; the check relates to pessimisations
the compiler might have to do and so it's perfectly sound to leave it
disabled.