Enable clang-tidy miscellaneous checks (#934)
This PR enables the "miscellaneous" set of `clang-tidy` checks. Adopting
these checks requires basically no changes to our code, but I've left
the following checks disabled:
- Disallowing recursion
- Disallowing unused parameters with names (e.g. this check prefers `T
f(A, B)` to `T f(A a, B b)` when `a` and `b` are unused).
- Using public data members in classes; we do this in a few places and I
think it's an acceptable decision.

As ever, happy to bikeshed but this shouldn't be a controversial one at
all.