Hotfix for byte escaping issues (#944)
The frontend test suite catches a bug introduced in #933, where byte
literal values that overflow a signed character are printed incorrectly.
I took a look at the best place to test this (i.e. whether it was worth
backporting the relevant test into the backend test suite), and it is
indeed in the frontend because of the way that we provide test cases in
each project.

I have run the downstream K test suite and verified that it passes with
this fix applied.

Fixes: https://github.com/runtimeverification/k/pull/3883