Enable fixes for memory leaks in bindings (#945)
This PR fixes a long-standing issue in the backend's binding code, which
is that the call to `initStaticObjects()` when the Python module is
_bound_ does not affect the process that actually runs the backend
interpreter. A consequence of this is that GMP is never told to use our
GC'ed memory allocators, and so any code using integer operations ends
up leaking lots of memory when it runs for a long time.

To fix Python code that uses the bindings:
* Call `init_static_objects` on application startup
* If an application _only_ does function evaluation or simplification,
you may also want to call `free_all_gc_memory` occasionally as the
backend will not do so autonomously unless a rewrite step is taken.

Relevant:
* #770
* #922 (has reproduction instructions)