Enable clang-tidy readability checks (#936)
This is another PR in my ongoing series to enable more clang-tidy checks
for the codebase. As previously, each commit in this PR fixes the issues
flagged by a single check, and can / should be reviewed independently.

As ever, I'm happy to bikeshed any of the specific changes in this PR.
Note that for
[`readability-function-cognitive-complexity`](https://github.com/runtimeverification/llvm-backend/pull/936/commits/c2b0de149776d1af54a519e81ecf53923dfd1ec1),
I've not actually refactored any existing code in the interest of
keeping this diff minimal, but the check will prevent us from
introducing new code that exceeds the complexity threshold.