Refactor subsort table code (#951)
The backend had duplicated logic in a couple of places to build the
subsort[^1] relation for a definition; this PR refactors that code to
live with `KOREDefinition` so that it can be more easily reused. I have
also taken the opportunity to pull some code out of the `AST.cpp`
monolith into a separate translation unit to make it easier to read.

No actual functionality is introduced here; it's just a refactoring to
move identical code inside a common method.

[^1]: Analogously for overloads as well