Interpret C++ exceptions in C bindings (#955)
This PR amends the interface of the simplification routines in the C
bindings to allow for error information to be safely returned to calling
clients without crashing the host process.

The changes are as follows:
- Introduce a new opaque `kore_error` type to the C bindings, with
appropriate bindings.
- **Breaking change** to the interfaces of `kore_simplify`,
`kore_simplify_bool` and `kore_simplify_binary` to take an optional
additional `kore_error` parameter. If a C++ exception is thrown during
these bindings, it will be caught and used to populate the error
parameter's fields. The booster will need to be updated as part of its K
release job down the line.
- Add a test that catches an error that imitates the original reporting
of this issue (invalid numeric value in the hooked arithmetic library).

Closes #953