Implement helper to collect sorts hooked to a particular implementation (#952)
The method implemented here collects all the sorts that are hooked to a
particular implementation; typically this will be a singleton set, but
it is possible in theory to have user code add additional sorts.

Motivated by @Scott-Guest's observation on #950 that it is not safe to
assume `SortBytes` is the only sort on which byte hooks can be called;
we instead need to compute the set of sorts greater than any sort hooked
to `BYTES.Bytes` (using code in #951).