Fix for possible silent failures in test suite (#959)
This PR fixes an issue whereby we could experience silent failures in
the lit test suite when mismatches occur between command outputs being
compared with `diff`; this was caused by a deficiency in the `one_line`
function in our lit configuration that meant only the _last_ command
executed could cause the test to fail.

Fortunately, no actual regressions were caused by this issue. Several
tests had slightly malformed output that could be fixed by hand
(newlines, extra spaces), and two tests (31, 32) experienced a
regression due to their reliance on non-zero exit codes. This regression
is fixed by adding a check that the `--binary-output` mode does not
print any messages; the actual behaviour of this line is checked by the
conversion on the subsequent line.

I have verified that the fix in this PR addresses the case where I first
reproduced this issue, and that there are no more silent failures in the
fully-verbose output from the test suite.