Precompute subsort, overload relations
Previously, we computed these relations once each for code generation
and pretty-printing. The implementation of the CoW mechanism will
require access to the relations in an inner loop during code generation,
and so we need to precompute them when a definition is loaded.

Because KOREDefinition objects are built piece-by-piece from the parser,
we hide its constructor and ensure that the necessary preprocessing step
is performed when a definition is loaded from disk.