Properly handle subsorting and subterms
Prior to this commit, the implementation would only work for variables
that appear explicitly on the LHS of a rewrite, and are directly of
static sort `SortBytes`. This is not sufficient, and breaks in the
following scenarios:

- Sorts that are not Bytes, but are hooked to the same implementation (I
  refer to this set of sorts as Bytes for convenience).
- Terms that are statically of a supersort of Bytes, and at runtime are
  terms of sort Bytes.
- Configurations that can syntactically contain a Bytes subterm, but do
  not explicitly mention such a variable in the LHS of a rule.

To properly handle these cases, we need to use the precomputed "sort
contains sort" relation to look up whether any subterm of an aliased
variable could contain a Bytes term, and to recurse through runtime
terms using a visitor in order to mark all such Bytes subobjects as
copy-on-write.