Add `--mutable-bytes` flag to enable possibly unsound optimizations (#962)
This PR adds a flag to `llvm-kompile` that enables semantics used by the
LLVM backend that are unsound in certain scenarios, but are useful for
performance (making objects of sort `Bytes` copy-on-write rather than
mutable).

The numerous edge cases from my initial attempt to fix this issue in
#950 are retained as test cases here.

Fixes #930 

Blocked until we have propagated
https://github.com/runtimeverification/k/pull/3934 appropriately to
downstream semantics.