update to python 3.10 (#966)
In order to support seamless integration between the llvm backend and
pyk on nix, we need to build the llvm backend with the same version of
python that we use at runtime. The version we use at runtime, according
to pyk's `pyproject.toml` file, is 3.10, so we update the version of
python we use in the nix derivation for the llvm backend.