Use FILE * instead of filenames when passing file paths to serialization and printing commands (#968)
Previously we were storing a filename in a global variable and then
passing it to a bunch of functions. Each of these functions then had to
call fopen and fclose, which was quite expensive. This PR changes the
behavior so instead we once open a file with fopen at the start of
rewriting, then call a bunch of functions which take a `FILE *` as an
argument, then call fclose once at the end. This gets much better
buffering behavior and reduces the number of syscalls considerably.

When measuring the performance of this code against a "sum to 10k" imp
program with proof hint generation enabled, we saw a time reduction from
3.5m to 2m.