Look up precomputed symbol names for {} symbols
If a symbol doesn't have any formal sort parameters (i.e. it isn't an
instantiation of a parametric symbol), we don't need to do the extra
work of parsing out its sort list and correctly special-casing for
injections. Instead, we can just look up its name and directly emit
those bytes into the serializer buffer.

The code here will be cleaned up in subsequent commits, but it
implements the core of the relevant change in this PR.