Allow partial function evaluation to be caught by bindings (#967)
This PR extends the backend's ability to recover from errors when
partial functions are evaluated in the context of a bindings library.
Rather then unconditionally crashing the entire host process, we throw
an exception that the bindings code can catch and translate to a C-ABI
error structure for the backend to deal with.

Previously, we implemented this behaviour for hooks
(https://github.com/runtimeverification/llvm-backend/pull/955); this PR
does the same for general partial function evaluation. The changes are
as follows:
* Refactor (as discussed previously) the implementation of
`finish_rewriting` from LLVM IR into C++.
* Add a new code-generation flag to enable the new error behaviour.
* Modify `llvm-kompile` to pass this flag when compiling a bindings
library.
* Read the flag to change behaviour in the appropriate places in the
runtime library.
* Add a test that the C bindings can safely try to evaluate undefined
partial functions and catch an error when the evaluation fails.

Fixes https://github.com/runtimeverification/llvm-backend/issues/925