support pyk form of axioms in splitTop
When pyk generates new KORE modules for use internal to its prover, it
creates rules whose structure is different from any rule which is
generated by the K frontend. As a result, splitTop does not know how to
handle such axioms. We thus need to extend splitTop to support this type
of rule in order to support parsing these modules as needed by the Maude
backend.