Implement type-safe attributes (#983)
This PR implements a more limited version of the type-safe attribute
infrastructure that we use in the K frontend; by doing so we can
construct a canonical list of the attributes that we _actually_ use in
the backend, and avoid working with string data when we don't need to
be.

The core of the change is a new `attribute_set` structure that _accepts_
attributes permissively from the parser, but only allows key-value
lookup via a limited set of whitelisted attribute keys. There are a
couple of places in the code that need to access the full set of
attributes, ignoring the whitelist:
* When pretty-printing via iteration over the set of stored attributes
* When returning the set of attributes to Pyk via the Python bindings

Both of these code paths have comments to make sure they are used
cautiously.

The bulk of the changes are 1:1 and can be reviewed pretty easily in the
split diff; there's only a small API change in code that consumes
attributes in the backend.

There is a small papercut in that the syntax to reference an attribute
key is quite verbose (`attribute_set::key::name`); when we upgrade to
C++20 in the near future we get access to `using enum
attribute_set::key;` which solves this problem and makes the code nicer
to read.

I've tested the change upstream against the K integration test suite.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>